gauge fixed when no data is available and reallocate function
